ripples_api/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
torchvision
pillow
numpy
matplotlib
